fix: :bug: avoid removing disallowed file windows

+++ continuedev/src/continuedev/server/meilisearch_server.py
@@ -69,7 +69,7 @@ async def ensure_meilisearch_installed() -> bool:
         else:
             non_existing_paths.add(path)
 
-    if len(non_existing_paths) > 0:
+    if len(non_existing_paths) > 0 and os.name != "nt":
         # Clear the meilisearch binary
         if meilisearchPath in existing_paths:
             os.remove(meilisearchPath)
@@ -134,5 +134,5 @@ async def start_meilisearch():
             stderr=subprocess.STDOUT,
             close_fds=True,
             start_new_session=True,
-            shell=True
+            shell=True,
         )
